Fix crash in openEditor if SIGINT was delivered
nvi for example sets ISIG and traps SIGINT.
Without this patch, this would propagate to all processes in the same
process group and kill them. (It still does, but we set a signalHandler
to ignore that.)

Still not perfect, because for some reason we receive WIFSIGNALED even
if the signal did not actually kill the editor. For now, we just treat
this as a success.