regex: copy after compiling
Instead of the broken attempt at making regexes zero-copy (it copied
anyway), copy once and forget about it.

(There have been way too many problems with the destructor approach,
including the latest one where the GC would happily zero out our
regexes if they were in a sequence.

Maybe we can make this work once we switched to ORC. For now, it's
not worth the trouble.)