buffer: remove BufferSource
Aside from being a wrapper of Request, it was just storing the -I
charset, except even that didn't actually work. Whoops.

This fixes -I effectively not doing anything; now it's a forced override
that even disables BOM sniffing. (If the user wants to decode a file
using a certain encoding, it seems wise to assume that they really
meant it.)