ua.css: use ANSI yellow for links
This was my original intention, but there wasn't a way to do it until
now. The difference is that this respects user-configured color values.

(Now that I think of it, it may be better to automatically detect
prefers-color-scheme based on the default background color, and then use
blue for bright backgrounds and yellow for dark backgrounds. Hmm.)