Fix tab size bug on double tabs
It is in fact quite simple to calculate: charwidth is the width
of printing characters until now, and whitespacenum the number of
(non-flushed) spaces. So we just have to subtract this from the target
width to get the number of spaces missing from the next tab stop.

(Of course, it gets much harder to understand when the whole thing is
formatted as a convoluted multi-line equation...)