http: fix broken early hint handling
The empty string comparison here was in fact pointless; in cw-out.c,
libcurl only calls cwb (which is curlWriteHeader in this case) if blen
is not 0, so the string will never be empty. (Instead, it is expected
to be \r\n; I've added \n too since CGI can already parse headers like
that.)

Normally it still worked because we just passed through the line to
cgi.nim.  However, it choked horribly on HTTP/2 early hints.