dom: use pointers instead of numerical ids for collections
The "id" scheme had obvious problems when multiple documents existed.

Originally it was needed because the old hacky integration with QuickJS
would occasionally result in objects being moved to other addresses.

This has been fixed long ago when I decided to vendor in a fork, so we
can just use pointers as ids unique to the entire process.