sandbox: allow syscalls for epoll Nim selectors
This fixes setTimeout/setInterval causing crashes.

Note: timerfd_gettime is not actually used by Nim right now.
However, it seems like a good idea to add it to the set in case a
future Nim version needs it, as it does no harm.

We still do not allow signalfd, because it would let rogue buffers
override our SIGSYS handler.  (Not sure if this really matters, but
we don't need it for now anyway.)