base64: rewrite atob
Turns out std/base64's `decode' is broken: atob(" ") would panic.  So we
no longer use that.

Basic testing indicates that the new version is closer to the standard-
mandated behavior than the old one was.  OTOH I assume it's somewhat
slower, but that can be improved later if it proves to be a bottleneck.