layout: resize over-constrained table columns
If the total specified column width is larger than the table's allowed
width, we now resize them proportionally to the specified width.

This is quite important because many tables set the width to e.g. 50%
for "give me half of the table", instead of its true meaning "give me
half of the page".