layout: fix BFC positioning in presence of floats
Previously, the following printed "2" instead of "12":

<div style=float:left>1</div><div style=display:flow-root>2</div>

This fixes the above problem in the laziest possible way: we relayout once
in the smallest possible space the BFC fits in when floats exist. See the
comment in the code for details.

As a nice bonus, this also fixes tables overlapping with floated boxes,
by pretending that they establish a BFC.