sixel: dithering
uses floyd-steinberg - I originally tried atkinson, but liked the
results with fs a bit more.

also, I got rid of the half hearted attempt to skip tree traversals for
color lookups, as it performs horribly with dithering. it *did* work if
I set Y as the key, but it still felt wrong - well, as it turns out,
octree traversal is faster anyway, so just do that.

it works out nicely because we can fill in holes (from dithering) with
a linear search for the nearest match as we go.